rename Guild and Member
