+++ src/Guild.js
@@ -1,11 +1,11 @@
 const RateLimiter = require('./RateLimiter');
 const Discord = require('discord.js');
 
-class GuildHandler {
+class Guild {
 
     /**
-     * GuildHandler Cache
-     * @type {module:"discord.js".Collection<module:"discord.js".Snowflake, GuildHandler>}
+     * Guild Cache
+     * @type {module:"discord.js".Collection<module:"discord.js".Snowflake, Guild>}
      */
     static #cache = new Discord.Collection();
 
@@ -66,17 +66,17 @@ class GuildHandler {
     /**
      * get a guild from cache or create a new one
      * @param {module:"discord.js".Guild}   guild
-     * @return {GuildHandler}
+     * @return {Guild}
      */
     static get(guild) {
         if (this.#cache.has(guild.id))
             return this.#cache.get(guild.id)
         else {
-            const newGuild = new GuildHandler(guild);
+            const newGuild = new Guild(guild);
             this.#cache.set(guild.id, newGuild);
             return newGuild;
         }
     }
 }
 
-module.exports = GuildHandler;
+module.exports = Guild;
